+++ services/apisix/scripts/initialization/00-set-oidc-plugin.curl.sh
@@ -10,8 +10,8 @@ JSON_CONTENT=$(cat <<EOF
   "desc": "Avenirs OpenID Connect authentication configuration",
   "plugins": {
     "openid-connect": {
-      "client_id": "$APISIX_OIDC_CLIENT_ID",
-      "client_secret": "$APISIX_OIDC_CLIENT_SECRET",
+      "client_id": "$SEC_APISIX_OIDC_CLIENT_ID",
+      "client_secret": "$SEC_APISIX_OIDC_CLIENT_SECRET",
       "discovery": "https://avenirs-apache/cas/oidc/.well-known",
       "introspection_endpoint": "https://avenirs-apache/cas/oidc/introspect",
       "redirect_uri": "https?://avenirs-apache/node-api/cas-auth-callback(.*)",
@@ -27,7 +27,7 @@ JSON_CONTENT=$(cat <<EOF
 EOF
 )
 
-curl -H "X-API-KEY: $APISIX_ADMIN_KEY" -i "$END_POINT" -X PUT -d "$JSON_CONTENT"
+curl -H "X-API-KEY: $SEC_APISIX_ADMIN_KEY" -i "$END_POINT" -X PUT -d "$JSON_CONTENT"
 
 
 echo -ne "\n\n"